# Replace some instances of `Image` with nibabel equivalent to speed up script