<!-- Hi, and thank you for submitting a Pull Request! The checklist below is a brief summary of steps found in the NeuroPoly Contributing Guidelines, which can be found here: https://www.neuro.polymtl.ca/software/contributing. 
-->

## Checklist

#### GitHub

- [X] I've given this PR a concise, self-descriptive, and meaningful title
- [X] I've linked relevant issues in the PR body
- [X] I've applied [the relevant labels](https://www.neuro.polymtl.ca/software/contributing#pr_labels) to this PR
- [ ] I've applied a [release milestone](https://github.com/neuropoly/spinalcordtoolbox/milestones) (major, minor, patch) in line with [Semantic Versioning guidelines](https://github.com/neuropoly/spinalcordtoolbox/wiki/Misc%3A-Creating-a-new-release#convention-for-naming-releases) 
- [ ] I've assigned a reviewer

<!-- For the title, please observe the following rules:
	- Provide a concise and self-descriptive title
	- Do not include the applicable issue number in the title, do it in the PR body
	- If the PR is not ready for review, convert it to a draft.
-->

#### PR contents

- [ ] I've consulted [SCT's internal developer documentation](https://github.com/neuropoly/spinalcordtoolbox/wiki) to ensure my contribution is in line with any relevant design decisions
- [ ] I've added [relevant tests](https://github.com/neuropoly/spinalcordtoolbox/wiki/Programming%3A-Tests) for my contribution
- [ ] I've updated the [relevant documentation](https://github.com/neuropoly/spinalcordtoolbox/wiki/Programming%3A-Documentation) for my changes, including argparse descriptions, docstrings, and ReadTheDocs tutorial pages

## Description
<!-- describe what the PR is about. Explain the approach and possible drawbacks.It's ok to repeat some text from the related issue. -->

Following investigation in #3271, this PR removes some `Image` calls in `sct_apply_transfo` and replaces them with their `nibabel` equivalents to see if this would speed up the script. 

To test the improvement, I timed the command recommended by @jcohenadad in https://github.com/neuropoly/spinalcordtoolbox/pull/3272#issuecomment-805888050:

```
sct_warp_template -d sct_example_data/t2/t2.nii.gz -w sct_example_data/t2/warp_template2anat.nii.gz -a 1 -s 1
```

Here are the logs/results:
* [`499.36s user 75.58s system 163% cpu 5:51.54 total`](https://github.com/neuropoly/spinalcordtoolbox/files/6201003/sct_warp_template-master-t2.txt) (master)
* [`392.10s user 39.45s system 173% cpu 4:09.17 total` ](https://github.com/neuropoly/spinalcordtoolbox/files/6200933/sct_warp_template-jn-3271-speed-up-sct_apply_transfo.txt) (this PR)

### Future work

* The [new call to `nib.save`](https://github.com/neuropoly/spinalcordtoolbox/blob/484771e35eb1d8f1906e0abfa57d80d99a8e7d7c/spinalcordtoolbox/scripts/sct_apply_transfo.py#L301) (for copying the xforms) is an improvement, but still a relatively costly operation (~1.5s) which adds up to ~1.5m across the `sct_warp_template` call. But, this seems like a necessary cost for fixing the affines.
* There are other `Image` calls in `sct_apply_transfo` (and all throughout SCT). Adapting everything would be a very large undertaking. My goal with this PR was to just to quickly experiment with working with nibabel images directly. But, I'd like to explore other ways of improving the performance of the `Image` class, too.

## Linked issues
<!-- If the PR fixes any issues, indicate it here with issue-closing keywords: e.g. Resolves #XX, Fixes #XX, Addresses #XX. Note that if you want multiple issues to be autoclosed on PR merge, you must use the issue-closing verb before each relevant issue: e.g. Resolves #1, Resolves #2 -->

Partially addresses #3271.
Follow-up to #3272.