@joshuacwnewton 2021-03-24
Running into `Bus error (core dumped)` again. I'll debug this and then mark this as ready for review.