@joshuacwnewton 2021-05-01
I'm going to close this for now. 

On reflection, this feels like a bandaid solution that doesn't actually address the underlying problem (#3271). I would rather try to fix our `Image` class than avoiding `Image` like I did here in this PR .